Revert "added xbox support"
This reverts commit aad6b2d9fd243bbd8d2eef3ab457493b80f5e963.
