+++ src/main/java/org/usfirst/frc/team3414/actuators/Climber.java
@@ -18,7 +18,6 @@ import org.usfirst.frc.team3414.diagnostic.LEDColor;
 import org.usfirst.frc.team3414.teleop.Teleop;
 
 import edu.wpi.first.wpilibj.DigitalInput;
-import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
@@ -85,7 +84,7 @@ public class Climber {
         double topClimbTime = 8;
         double midClimbTime = 4;
         double midToTopClimbTime = 4;
-        while (getEscapeButton() && DriverStation.getInstance().isEnabled()) {
+        while (getEscapeButton()) {
             updateDashboard(stage);
             if (stage == 1) {
                 LED.set(LEDColor.RED); // Sets the LEDs
@@ -176,19 +175,6 @@ public class Climber {
         }
         stop();
 
-    }
-    public void holdClimb(int target, boolean driverControl){
-        rear(rearMotor);
-        front(frontMotor);
-        while (getEscapeButton() && DriverStation.getInstance().isEnabled()) {
-       
-        setFront(target);
-        setRear(target);
-        if(driverControl){
-            middleMotor.set(ControlMode.PercentOutput, (Teleop.getInstance().getLeftJoystick().getY()+Teleop.getInstance().getRightJoystick().getY())/2);
-        }
-        }
-        stop();
     }
 
     // TALON CONFIG
