+++ src/main/java/org/usfirst/frc/team3414/teleop/Controller.java
@@ -3,79 +3,90 @@ package org.usfirst.frc.team3414.teleop;
 import org.usfirst.frc.team3414.config.Config;
 
 import edu.wpi.first.wpilibj.Joystick;
-import edu.wpi.first.wpilibj.XboxController;
-import edu.wpi.first.wpilibj.GenericHID.Hand;
 
 public class Controller {
 	//Normal Controllers with Normal Mappings. There is a difference.
 	
-    F310 pad;
-    XboxController xbox;
-	public  Controller(){
-        pad = new F310(Config.CONTROLLER_CHANNEL);
-        xbox = new XboxController(Config.XBOX_CONTROLLER_CHANNEL);
+	Joystick pad;
+	public Controller(int channel){
+		pad = new Joystick(channel);
 		}
 		public void closeOut() {
 			pad = null;
 		}
 	public boolean getAButton() {
-		return pad.getAButton() || xbox.getAButton();
+		return pad.getRawButton(2);
 
 	}
 
 	public boolean getBButton() {
-		return pad.getBButton() || xbox.getAButton();
+		return pad.getRawButton(3);
 
 	}
 
 	public boolean getXButton() {
-		return pad.getXButton() || xbox.getXButton();
+		return pad.getRawButton(1);
 
 	}
 
 	public boolean getYButton() {
-        return pad.getYButton() || xbox.getYButton();
+		return pad.getRawButton(4);
+
 	}
 
 	public boolean getLBButton() {
-        return pad.getLBButton() || xbox.getBumper(Hand.kLeft);
+		return pad.getRawButton(5);
+
 	}
 
 	public boolean getRBButton() {
-		return pad.getRBButton() || xbox.getBumper(Hand.kRight);
+		return pad.getRawButton(6);
 	}
 
 	public boolean getLT() {
-		return pad.getLT() || xbox.getTriggerAxis(Hand.kLeft) > .25;
+		return pad.getRawButton(7);
 
 	}
 
 	public boolean getRT() {
-		return pad.getRT() || xbox.getTriggerAxis(Hand.kRight) > .25;
+		return pad.getRawButton(8);
 
 	}
 	public boolean getLSButton(){
-        return pad.getLSButton() || xbox.getStickButton(Hand.kLeft);
-    	}
+		return pad.getRawButton(11);
+	}
 	public boolean getRSButton(){
-        return pad.getRSButton() || xbox.getStickButton(Hand.kRight);
+		return pad.getRawButton(12);
 	}
 
 	public double getPov() {
-		return pad.getPov();
-    }
-    public double getXPov(){
-        return xbox.getPOV();
-    }	
-	
+		return pad.getPOV();
+	}
+	public boolean isPovPositive(){
+		if((135 >= pad.getPOV()) && (pad.getPOV() <= 45)){
+			return true;
+		}
+		else{
+			return false;
+		}		
+	}
+	public boolean isPovNegative(){
+		if((225 >= pad.getPOV()) && (pad.getPOV() <= 315)){
+			return true;
+		}
+		else{
+			return false;
+		}		
+	}
 private static Controller instance;
 
 public static Controller getInstance()
 		{
 			if(instance == null)
 			{
-				instance = new Controller();
-            }
+				instance = new Controller(Config.CONTROLLER_CHANNEL);
+			}
+			
 			return instance;
 			
 	}
